refactor(backend): reduce cyclomatic complexity — extract billing, dedup, and filter pipeline modules (#1781)
Extract services/billing.py into services/billing/ subpackage:
  - checkout.py: Intel Report Stripe Checkout creation
  - subscription.py: billing period management + next billing date
  - __init__.py: facade re-exporting all public symbols

Extract consolidation/dedup.py into consolidation/dedup/ subpackage:
  - layer1_exact.py: source_id deduplication
  - layer2_key.py: dedup_key exact dedup + merge-enrichment
  - layer3_fuzzy.py: Jaccard fuzzy deduplication
  - layer4_process.py: process-number deduplication
  - layer5_title.py: title-prefix deduplication
  - engine.py: DeduplicationEngine orchestrator with backward-compat wrappers
  - _helpers.py: shared tokenization/Jaccard utilities

Extract filter/pipeline.py LLM-heavy blocks into filter/stages/:
  - llm_zero_match.py: GTM-FIX-028 zero-match LLM classification
  - recovery.py: FLUXO 2 anti-false-negative recovery pipeline
  - Main pipeline calls extracted functions as thin orchestrator

Backward compat: all original import paths preserved via facade re-exports.
Zero regressions: 327 tests pass, 0 failures.
