+++ backend/consolidation/dedup/__init__.py
@@ -1,0 +1,1 @@
+from consolidation.dedup.engine import DeduplicationEngine  # noqa: F401
