+++ backend/filter/pipeline.py
@@ -796,263 +796,15 @@ def aplicar_todos_filtros(
     # ========================================================================
     # Instead of auto-rejecting bids with 0 keyword matches, collect them
     # and send to LLM for sector-aware classification.
-    resultado_llm_zero: List[dict] = []
-    stats["llm_zero_match_calls"] = 0
-    stats["llm_zero_match_aprovadas"] = 0
-    stats["llm_zero_match_rejeitadas"] = 0
-    stats["llm_zero_match_skipped_short"] = 0
-    # STORY-354 AC9: count LLM fallbacks that returned pending_review=True
-    stats["pending_review_count"] = 0
-
-    # STORY-267 AC2: When custom_terms present + TERM_SEARCH_LLM_AWARE, use term-aware prompt
-    _use_term_prompt_zm = False
-    if custom_terms:
-        from config import get_feature_flag as _gff_ac2
-        _use_term_prompt_zm = _gff_ac2("TERM_SEARCH_LLM_AWARE")
-
-    # ISSUE-017: Also run zero-match LLM when custom_terms present (no sector needed)
-    if setor or custom_terms:
-        # Collect bids that were rejected by keyword gate (in resultado_valor but not in resultado_keyword)
-        keyword_approved_ids = {id(lic) for lic in resultado_keyword}
-        zero_match_pool: List[dict] = []
-        for lic in resultado_valor:
-            if id(lic) not in keyword_approved_ids:
-                objeto = lic.get("objetoCompra", "")
-                # AC3: Skip bids with objeto < 20 chars (PCP short resumo, insufficient signal)
-                if len(objeto) < 20:
-                    stats["llm_zero_match_skipped_short"] += 1
-                    logger.debug(
-                        f"LLM zero_match: SKIP (objeto < 20 chars) objeto={objeto!r}"
-                    )
-                    continue
-                zero_match_pool.append(lic)
-
-        # ISSUE-029: Pre-filter bids whose objetoCompra contains sector negative_keywords.
-        # Ported from filter_llm.py — the extracted version was NOT on the production path.
-        if zero_match_pool and setor:
-            _sector_negative_kws: list = []
-            try:
-                from sectors import get_sector as _get_sector_neg
-                _neg_sec = _get_sector_neg(setor)
-                _sector_negative_kws = [kw.lower() for kw in getattr(_neg_sec, "negative_keywords", [])]
-            except Exception:
-                pass
-
-            if _sector_negative_kws:
-                _neg_filtered = []
-                _neg_rejected = 0
-                for lic in zero_match_pool:
-                    _obj_lower = (lic.get("objetoCompra") or "").lower()
-                    if any(neg_kw in _obj_lower for neg_kw in _sector_negative_kws):
-                        _neg_rejected += 1
-                        logger.debug(
-                            f"LLM zero_match: PRE-FILTER (negative_keyword) "
-                            f"objeto={lic.get('objetoCompra', '')[:80]}"
-                        )
-                    else:
-                        _neg_filtered.append(lic)
-                if _neg_rejected > 0:
-                    logger.info(
-                        f"[ISSUE-029] Zero-match negative_keyword pre-filter: "
-                        f"removed {_neg_rejected}/{len(zero_match_pool)} bids before LLM"
-                    )
-                    stats["llm_zero_match_neg_prefilter"] = _neg_rejected
-                zero_match_pool = _neg_filtered
-
-        if zero_match_pool:
-            from llm_arbiter import classify_contract_primary_match as _classify_zm
-            from sectors import get_sector as _get_sector_zm
-
-            if setor:
-                try:
-                    setor_config_zm = _get_sector_zm(setor)
-                    setor_name_zm = setor_config_zm.name
-                except (KeyError, Exception):
-                    setor_name_zm = setor
-            elif custom_terms:
-                # ISSUE-017: No sector — use custom_terms as context for LLM
-                setor_name_zm = ", ".join(custom_terms[:3])
-
-            # AC6: Concurrent LLM calls with max 10 threads (equivalent to Semaphore(10))
-            def _classify_one(lic_item: dict) -> tuple[dict, dict]:
-                obj = lic_item.get("objetoCompra", "")
-                # STORY-328 AC14: Use stripped text for LLM classification
-                obj = _strip_org_context(obj)
-                val = lic_item.get("valorTotalEstimado") or lic_item.get("valorEstimado") or 0
-                if isinstance(val, str):
-                    try:
-                        val = float(val.replace(".", "").replace(",", "."))
-                    except ValueError:
-                        val = 0.0
-                else:
-                    val = float(val) if val else 0.0
-                # STORY-267 AC2: Use term-aware prompt when custom_terms present
-                if _use_term_prompt_zm and custom_terms:
-                    result = _classify_zm(
-                        objeto=obj,
-                        valor=val,
-                        setor_name=None,
-                        termos_busca=custom_terms,
-                        prompt_level="zero_match",
-                        setor_id=None,
-                    )
-                else:
-                    result = _classify_zm(
-                        objeto=obj,
-                        valor=val,
-                        setor_name=setor_name_zm,
-                        prompt_level="zero_match",
-                        setor_id=setor,
-                    )
-                return lic_item, result
-
-            _llm_total = len(zero_match_pool)
-            _llm_completed = 0
-
-            # STORY-4.1 (TD-SYS-014): Replaced ThreadPoolExecutor(max_workers=10) with
-            # asyncio.gather bounded by LLM_MAX_CONCURRENT (default 50). The inner
-            # _classify_one remains sync; async_runtime.run_bounded_in_thread executes
-            # it in a worker thread under the shared semaphore + Prometheus gauge.
-            import asyncio as _asyncio_zm
-
-            from llm_arbiter.async_runtime import (
-                gather_classifications as _gather_zm,
-            )
-            from llm_arbiter.async_runtime import (
-                unwrap_result as _unwrap_zm,
-            )
-
-            def _on_zm_progress(done: int, total: int, phase: str) -> None:
-                nonlocal _llm_completed
-                _llm_completed = done
-                stats["llm_zero_match_calls"] += 1
-                if on_progress:
-                    on_progress(done, total, phase)
-
-            _paired_results = _asyncio_zm.run(
-                _gather_zm(
-                    _classify_one,
-                    list(zero_match_pool),
-                    call_type="zero_match",
-                    on_progress=_on_zm_progress,
-                )
-            )
-
-            class _ResolvedFuture:
-                def __init__(self, value): self._v = value
-                def result(self): return _unwrap_zm(self._v)
-
-            futures = {_ResolvedFuture(r): None for r in _paired_results}
-
-            for future in futures:
-                try:
-                    lic_item, llm_result = future.result()
-                    is_relevant = llm_result.get("is_primary", False) if isinstance(llm_result, dict) else llm_result
-                    # ISSUE-017: Post-LLM gate — reject if custom terms not in text
-                    if is_relevant and custom_terms:
-                        obj_norm = normalize_text(lic_item.get("objetoCompra", ""))
-                        has_term_evidence = any(
-                            normalize_text(term) in obj_norm
-                            for term in custom_terms
-                        )
-                        if not has_term_evidence:
-                            is_relevant = False
-                            logger.debug(
-                                f"LLM zero_match: OVERRIDE to NO — no custom term "
-                                f"found in text. terms={custom_terms}, "
-                                f"objeto={lic_item.get('objetoCompra', '')[:80]}"
-                            )
-                    if is_relevant:
-                        stats["llm_zero_match_aprovadas"] += 1
-                        # STORY-267 AC16: Track term search metrics
-                        if custom_terms:
-                            from metrics import TERM_SEARCH_LLM_ACCEPTS
-                            TERM_SEARCH_LLM_ACCEPTS.labels(zone="zero_match").inc()
-                        # AC8: Tag relevance source
-                        lic_item["_relevance_source"] = "llm_zero_match"
-                        lic_item["_term_density"] = 0.0
-                        lic_item["_matched_terms"] = []
-                        # D-02 AC4: Confidence capped at 70 for zero-match
-                        if isinstance(llm_result, dict):
-                            raw_conf = llm_result.get("confidence", 60)
-                            lic_item["_confidence_score"] = min(raw_conf, 70)
-                            lic_item["_llm_evidence"] = llm_result.get("evidence", [])
-                        else:
-                            lic_item["_confidence_score"] = 60
-                            lic_item["_llm_evidence"] = []
-                        resultado_llm_zero.append(lic_item)
-                        logger.debug(
-                            f"LLM zero_match: ACCEPT conf={lic_item.get('_confidence_score')} "
-                            f"objeto={lic_item.get('objetoCompra', '')[:80]}"
-                        )
-                    else:
-                        # STORY-354 AC9: when LLM fails with LLM_FALLBACK_PENDING_ENABLED
-                        # the arbiter tags ``pending_review=True``. The filter-level contract
-                        # (AC19, pinned in tests/test_llm_zero_match.py) stays REJECT — the
-                        # bid is still counted in llm_zero_match_rejeitadas and NOT forwarded
-                        # to aprovadas. Downstream (pipeline/stages/generate.py) consumes
-                        # ``stats["pending_review_count"]`` to store the bid in Redis and
-                        # enqueue the reclassify job — that's where the PENDING_REVIEW
-                        # response metadata is attached, not here.
-                        if isinstance(llm_result, dict) and llm_result.get("pending_review"):
-                            stats["pending_review_count"] += 1
-                            lic_item["_pending_review"] = True
-                            lic_item["_pending_review_reason"] = llm_result.get(
-                                "rejection_reason", "LLM unavailable"
-                            )
-                            logger.debug(
-                                f"LLM zero_match: PENDING_REVIEW tagged (REJECT) "
-                                f"objeto={lic_item.get('objetoCompra', '')[:80]}"
-                            )
-                        stats["llm_zero_match_rejeitadas"] += 1
-                        # STORY-267 AC16: Track term search metrics
-                        if custom_terms:
-                            from metrics import TERM_SEARCH_LLM_REJECTS
-                            TERM_SEARCH_LLM_REJECTS.labels(zone="zero_match").inc()
-                        # D-02 AC6: Store rejection reason for audit
-                        if isinstance(llm_result, dict):
-                            lic_item["_llm_rejection_reason"] = llm_result.get("rejection_reason", "")
-                        logger.debug(
-                            f"LLM zero_match: REJECT objeto={lic_item.get('objetoCompra', '')[:80]}"
-                        )
-                except Exception as e:
-                    # AC9: Fallback on LLM failure = REJECT
-                    stats["llm_zero_match_rejeitadas"] += 1
-                    logger.error(f"LLM zero_match: FAILED (REJECT fallback): {e}")
-
-            # ISSUE-029: Acceptance ratio circuit breaker — narrow sectors cap acceptance.
-            # If LLM accepted too many zero-match bids, demote all to pending_review
-            # to prevent false positive flood (e.g. vestuario cap = 10%).
-            if resultado_llm_zero and setor:
-                _cb_threshold = 0.30  # Default 30%
-                try:
-                    _sec_cfg_cb = _get_sector_zm(setor)
-                    if hasattr(_sec_cfg_cb, "zero_match_acceptance_cap") and _sec_cfg_cb.zero_match_acceptance_cap is not None:
-                        _cb_threshold = _sec_cfg_cb.zero_match_acceptance_cap
-                except Exception:
-                    pass
-
-                _total_classified = stats["llm_zero_match_aprovadas"] + stats["llm_zero_match_rejeitadas"]
-                if _total_classified > 0 and stats["llm_zero_match_aprovadas"] / _total_classified > _cb_threshold:
-                    _accept_ratio = stats["llm_zero_match_aprovadas"] / _total_classified
-                    _demoted = len(resultado_llm_zero)
-                    logger.warning(
-                        f"[ISSUE-029] Zero-match acceptance ratio {_accept_ratio:.1%} exceeds "
-                        f"{_cb_threshold:.0%} for setor={setor!r} — demoting {_demoted} to pending_review"
-                    )
-                    for _lic in resultado_llm_zero:
-                        _lic["_relevance_source"] = "pending_review"
-                        _lic["_pending_review"] = True
-                        _lic["_pending_review_reason"] = "zero_match_high_acceptance_ratio"
-                    resultado_llm_zero = []
-
-            logger.info(
-                f"GTM-FIX-028 LLM Zero Match: "
-                f"{stats['llm_zero_match_calls']} calls, "
-                f"{stats['llm_zero_match_aprovadas']} approved, "
-                f"{stats['llm_zero_match_rejeitadas']} rejected, "
-                f"{stats['llm_zero_match_skipped_short']} skipped (short)"
-            )
+    from filter.stages.llm_zero_match import run_llm_zero_match
+    resultado_llm_zero, stats = run_llm_zero_match(
+        resultado_valor=resultado_valor,
+        resultado_keyword=resultado_keyword,
+        setor=setor,
+        custom_terms=custom_terms,
+        on_progress=on_progress,
+        stats=stats,
+    )
 
     # ========================================================================
     # GTM-RESILIENCE-D01: Camada 1C — Item Inspection for Gray Zone (0-5%)
@@ -1657,257 +1409,15 @@ def aplicar_todos_filtros(
     # ========================================================================
     # STORY-179 FLUXO 2: Anti-False Negative Recovery Pipeline
     # ========================================================================
-    # Recover contracts that were incorrectly rejected by keyword filters.
-    # This happens when:
-    # 1. Exclusion keywords reject legitimate contracts (Camada 1B)
-    # 2. Synonym near-misses not covered by keyword set (Camada 2B)
-    # 3. LLM recovery for ambiguous rejections (Camada 3B)
-    # 4. Zero results relaxation (Camada 4)
-
-    # Initialize FLUXO 2 stats
-    stats["recuperadas_exclusion_recovery"] = 0
-    stats["aprovadas_synonym_match"] = 0
-    stats["synonyms_auto_approved"] = 0  # CRIT-FLT-006 AC3
-    stats["recuperadas_llm_fn"] = 0
-    stats["recuperadas_zero_results"] = 0
-    stats["llm_arbiter_calls_fn_flow"] = 0
-    stats["zero_results_relaxation_triggered"] = False
-
-    # GTM-FIX-028 AC10: When LLM zero-match is enabled, skip FLUXO 2 to avoid
-    # double-classification (zero-match bids already went through LLM)
-    _skip_fluxo_2 = stats.get("llm_zero_match_calls", 0) > 0
-    if _skip_fluxo_2:
-        logger.info(
-            "GTM-FIX-028 AC10: FLUXO 2 DISABLED — LLM zero-match already classified "
-            f"{stats['llm_zero_match_calls']} bids"
-        )
-
-    # STORY-267 AC4+AC7: Term-aware recovery for custom_terms searches
-    _use_term_prompt_recovery = False
-    _use_term_synonyms = False
-    if custom_terms:
-        from config import get_feature_flag as _gff_ac4
-        _use_term_prompt_recovery = _gff_ac4("TERM_SEARCH_LLM_AWARE")
-        _use_term_synonyms = _gff_ac4("TERM_SEARCH_SYNONYMS")
-
-    # Run recovery when we have a sector OR when we have custom_terms with term synonyms enabled
-    _run_fluxo_2 = (setor or (_use_term_synonyms and custom_terms)) and not _skip_fluxo_2
-    if _run_fluxo_2:
-        from synonyms import find_synonym_matches, should_auto_approve_by_synonyms
-        if _use_term_synonyms and custom_terms:
-            from synonyms import find_term_synonym_matches
-        from sectors import get_sector as _get_sector
-
-        try:
-            setor_config = _get_sector(setor) if setor else None
-            setor_keywords = setor_config.keywords if setor_config else set()
-            setor_name = setor_config.name if setor_config else None
-
-            # Collect IDs of already-approved contracts to avoid duplicates
-            aprovadas_ids = {id(lic) for lic in aprovadas}
-
-            # ------------------------------------------------------------------
-            # Camada 1B + 2B: Re-scan contracts rejected at keyword stage
-            # ------------------------------------------------------------------
-            # We look at contracts that passed UF/status/value filters but were
-            # rejected by keyword matching (they exist in resultado_valor but
-            # not in resultado_keyword).
-            rejeitadas_keyword_pool: List[dict] = []
-            for lic in resultado_valor:
-                if id(lic) not in aprovadas_ids:
-                    rejeitadas_keyword_pool.append(lic)
-
-            logger.debug(
-                f"FLUXO 2 iniciando: {len(rejeitadas_keyword_pool)} contratos no pool de "
-                f"recuperação (rejeitados após filtros rápidos)"
-            )
-
-            recuperadas: List[dict] = []
-            llm_candidates_fn: List[dict] = []
-
-            for lic in rejeitadas_keyword_pool:
-                objeto = lic.get("objetoCompra", "")
-                if not objeto:
-                    continue
-
-                # STORY-267 AC7: Use term synonym matching when custom_terms present
-                if _use_term_synonyms and custom_terms:
-                    synonym_matches = find_term_synonym_matches(
-                        custom_terms=custom_terms,
-                        objeto=objeto,
-                    )
-                elif setor:
-                    # Camada 2B: Check synonym matches (original sector-based)
-                    synonym_matches = find_synonym_matches(
-                        objeto=objeto,
-                        setor_keywords=setor_keywords,
-                        setor_id=setor,
-                    )
-                else:
-                    synonym_matches = []
-
-                if not synonym_matches:
-                    continue  # No synonyms found, skip
-
-                # Check if auto-approve threshold is met (2+ synonyms)
-                if _use_term_synonyms and custom_terms:
-                    # For term searches, auto-approve with 2+ matches directly
-                    should_approve_flag = len(synonym_matches) >= 2
-                    matches = synonym_matches
-                else:
-                    should_approve_flag, matches = should_auto_approve_by_synonyms(
-                        objeto=objeto,
-                        setor_keywords=setor_keywords,
-                        setor_id=setor,
-                        min_synonyms=2,
-                    )
-
-                if should_approve_flag:
-                    # High confidence: 2+ distinct synonym matches → auto-approve
-                    stats["aprovadas_synonym_match"] += 1
-                    stats["synonyms_auto_approved"] += 1  # CRIT-FLT-006 AC3
-                    lic["_recovered_by"] = "synonym_auto_approve"
-                    lic["_synonym_matches"] = [
-                        f"{canon}≈{syn}" for canon, syn in matches
-                    ]
-                    recuperadas.append(lic)
-                    # STORY-267 AC16: Track term search synonym recoveries
-                    if custom_terms:
-                        from metrics import TERM_SEARCH_SYNONYM_RECOVERIES
-                        TERM_SEARCH_SYNONYM_RECOVERIES.inc()
-                    logger.debug(
-                        f"  Recuperada por sinônimos (auto): {matches} "
-                        f"objeto={objeto[:80]}"
-                    )
-                else:
-                    # 1 synonym match → ambiguous, send to LLM (Camada 3B)
-                    lic["_near_miss_synonyms"] = synonym_matches
-                    llm_candidates_fn.append(lic)
-
-            # ------------------------------------------------------------------
-            # Camada 3B: LLM Recovery for ambiguous synonym matches
-            # ------------------------------------------------------------------
-            if llm_candidates_fn:
-                from llm_arbiter import classify_contract_recovery
-
-                for lic in llm_candidates_fn:
-                    objeto = lic.get("objetoCompra", "")
-                    # STORY-328 AC14: Use stripped text for LLM classification
-                    objeto = _strip_org_context(objeto)
-                    valor = lic.get("valorTotalEstimado") or lic.get("valorEstimado") or 0
-                    if isinstance(valor, str):
-                        try:
-                            valor = float(valor.replace(".", "").replace(",", "."))
-                        except ValueError:
-                            valor = 0.0
-                    else:
-                        valor = float(valor) if valor else 0.0
-
-                    near_miss = lic.get("_near_miss_synonyms", [])
-                    near_miss_info = ", ".join(
-                        f"{canon}≈{syn}" for canon, syn in near_miss
-                    )
-
-                    stats["llm_arbiter_calls_fn_flow"] += 1
-                    # STORY-267 AC4: Use term-aware recovery when custom_terms present
-                    if _use_term_prompt_recovery and custom_terms:
-                        should_recover = classify_contract_recovery(
-                            objeto=objeto,
-                            valor=valor,
-                            rejection_reason="keyword_no_match + synonym_near_miss",
-                            termos_busca=custom_terms,
-                            near_miss_info=near_miss_info,
-                        )
-                    else:
-                        should_recover = classify_contract_recovery(
-                            objeto=objeto,
-                            valor=valor,
-                            rejection_reason="keyword_no_match + synonym_near_miss",
-                            setor_name=setor_name,
-                            near_miss_info=near_miss_info,
-                        )
-
-                    if should_recover:
-                        stats["recuperadas_llm_fn"] += 1
-                        lic["_recovered_by"] = "llm_recovery"
-                        lic["_synonym_matches"] = [
-                            f"{canon}≈{syn}" for canon, syn in near_miss
-                        ]
-                        recuperadas.append(lic)
-                        logger.debug(
-                            f"  Recuperada por LLM (FN flow): near_miss={near_miss_info} "
-                            f"objeto={objeto[:80]}"
-                        )
-
-            # Add recovered contracts to approved list
-            if recuperadas:
-                aprovadas.extend(recuperadas)
-                logger.info(
-                    f"FLUXO 2: {len(recuperadas)} contratos recuperados "
-                    f"(synonym_auto: {stats['aprovadas_synonym_match']}, "
-                    f"llm_recovery: {stats['recuperadas_llm_fn']})"
-                )
-
-            # ------------------------------------------------------------------
-            # Camada 4: Zero Results Relaxation
-            # ------------------------------------------------------------------
-            if len(aprovadas) == 0 and len(rejeitadas_keyword_pool) > 0:
-                stats["zero_results_relaxation_triggered"] = True
-                logger.info(
-                    "FLUXO 2 Camada 4: Zero results detected, attempting relaxation"
-                )
-
-                # Relaxation: accept any contract with at least 1 synonym match
-                for lic in rejeitadas_keyword_pool:
-                    if id(lic) in {id(r) for r in recuperadas}:
-                        continue  # Already recovered
-
-                    objeto = lic.get("objetoCompra", "")
-                    if not objeto:
-                        continue
-
-                    # STORY-267: Use term synonyms in relaxation when custom_terms present
-                    if _use_term_synonyms and custom_terms:
-                        synonym_matches = find_term_synonym_matches(
-                            custom_terms=custom_terms,
-                            objeto=objeto,
-                        )
-                    elif setor:
-                        synonym_matches = find_synonym_matches(
-                            objeto=objeto,
-                            setor_keywords=setor_keywords,
-                            setor_id=setor,
-                        )
-                    else:
-                        synonym_matches = []
-
-                    if synonym_matches:
-                        stats["recuperadas_zero_results"] += 1
-                        lic["_recovered_by"] = "zero_results_relaxation"
-                        lic["_synonym_matches"] = [
-                            f"{canon}≈{syn}" for canon, syn in synonym_matches
-                        ]
-                        aprovadas.append(lic)
-
-                if stats["recuperadas_zero_results"] > 0:
-                    logger.info(
-                        f"Camada 4 relaxation: recovered {stats['recuperadas_zero_results']} "
-                        f"contracts via single-synonym matching"
-                    )
-
-        except KeyError:
-            logger.warning(f"Setor '{setor}' não encontrado - pulando FLUXO 2")
-        except Exception as e:
-            logger.error(f"FLUXO 2 recovery failed: {e}", exc_info=True)
-
-    logger.debug(
-        f"FLUXO 2 resultado: "
-        f"synonym_auto={stats['aprovadas_synonym_match']}, "
-        f"llm_recovery={stats['recuperadas_llm_fn']}, "
-        f"zero_results={stats['recuperadas_zero_results']}, "
-        f"llm_calls_fn={stats['llm_arbiter_calls_fn_flow']}"
+    from filter.stages.recovery import run_fluxo_2_recovery
+    aprovadas = run_fluxo_2_recovery(
+        aprovadas=aprovadas,
+        resultado_valor=resultado_valor,
+        setor=setor,
+        custom_terms=custom_terms,
+        stats=stats,
     )
-    # ========================================================================
+
 
     stats["aprovadas"] = len(aprovadas)
